Add indexer connector inner engine build process
