+++ .gitmodules
@@ -1,3 +1,3 @@
-[submodule "src/shared_modules/http-request"]
-	path = src/shared_modules/http-request
-	url = https://github.com/wazuh/wazuh-http-request.git
+[submodule "src/engine/source/wazuh-http-request"]
+	path = src/engine/source/wazuh-http-request
+	url = https://github.com/wazuh/wazuh-http-request
