+++ src/engine/source/base/include/base/utils/stringUtils.hpp
@@ -145,6 +145,7 @@ std::string toSentenceCase(const std::string& str);
 bool isNumber(const std::string& str);
 
 bool replaceAll(std::string& data, std::string_view toSearch, std::string_view toReplace);
+bool haveUpperCaseCharacters(const std::string& str);
 
 } // namespace base::utils::string
