+++ src/engine/source/indexerconnector/qa/test_data/test_add_bulk_indexer_connector/event_delete.json
