+++ src/engine/source/indexerconnector/src/monitoring.hpp
@@ -13,9 +13,8 @@
 #define _MONITORING_HPP
 
 #include "HTTPRequest.hpp"
+#include "base/utils/stringUtils.hpp"
 #include "secureCommunication.hpp"
-#include "shared_modules/content_manager/src/dataDecoder.hpp"
-#include "stringHelper.h"
 #include <atomic>
 #include <chrono>
 #include <condition_variable>
@@ -30,35 +29,35 @@ constexpr auto INTERVAL = 60u;
 
 namespace HealthCheckRows
 {
-    enum Type : std::size_t
-    {
-        HEADER = 0,
-        DATA,
-        SIZE
-    };
+enum Type : std::size_t
+{
+    HEADER = 0,
+    DATA,
+    SIZE
+};
 };
 
 namespace HealthCheckColumns
 {
-    enum Type : std::size_t
-    {
-        EPOCH = 0,
-        TIMESTAMP,
-        CLUSTER,
-        STATUS,
-        NODE_TOTAL,
-        NODE_DATA,
-        DISCOVERED_CLUSTER_MANAGER,
-        SHARDS,
-        PRI,
-        RELO,
-        INIT,
-        UNASSIGN,
-        PENDING_TASKS,
-        MAX_TASK_WAIT_TIME,
-        ACTIVE_SHARDS_PERCENT,
-        SIZE
-    };
+enum Type : std::size_t
+{
+    EPOCH = 0,
+    TIMESTAMP,
+    CLUSTER,
+    STATUS,
+    NODE_TOTAL,
+    NODE_DATA,
+    DISCOVERED_CLUSTER_MANAGER,
+    SHARDS,
+    PRI,
+    RELO,
+    INIT,
+    UNASSIGN,
+    PENDING_TASKS,
+    MAX_TASK_WAIT_TIME,
+    ACTIVE_SHARDS_PERCENT,
+    SIZE
+};
 };
 
 /**
@@ -122,42 +121,43 @@ public:
                         {
                             // Get the health of the server.
                             HTTPRequest::instance().get(
-                                HttpURL(value.first + "/_cat/health?v"),
-                                [&](std::string response)
-                                {
-                                    // Remove the tabs and double spaces.
-                                    Utils::replaceAll(response, "\t", " ");
-                                    Utils::replaceAll(response, "  ", " ");
-                                    // Split the response by rows.
-                                    const auto rows {Utils::split(response, '\n')};
+                                {HttpURL(value.first + "/_cat/health?v"), {}, secureCommunication, headers},
+                                {[&](std::string response)
+                                 {
+                                     // Remove the tabs and double spaces.
+                                     base::utils::string::replaceAll(response, "\t", " ");
+                                     base::utils::string::replaceAll(response, "  ", " ");
+                                     // Split the response by rows.
+                                     const auto rows {base::utils::string::split(response, '\n')};
 
-                                    // Check if the response has the expected number of rows.
-                                    if (HealthCheckRows::SIZE == rows.size())
-                                    {
-                                        // Split the data row by spaces.
-                                        const auto fields {Utils::split(rows.at(HealthCheckRows::DATA), ' ')};
+                                     // Check if the response has the expected number of rows.
+                                     if (HealthCheckRows::SIZE == rows.size())
+                                     {
+                                         // Split the data row by spaces.
+                                         const auto fields {
+                                             base::utils::string::split(rows.at(HealthCheckRows::DATA), ' ')};
 
-                                        // Check if the response has the expected number of columns and if the status is
-                                        // green.
-                                        if (fields.size() == HealthCheckColumns::SIZE &&
-                                            fields.at(HealthCheckColumns::STATUS).compare("green") == 0)
-                                        {
-                                            value.second = true;
-                                        }
-                                        else
-                                        {
-                                            value.second = false;
-                                        }
-                                    }
-                                    else
-                                    {
-                                        value.second = false; // LCOV_EXCL_LINE
-                                    }
-                                },
-                                [&](const std::string& error, const long /*statusCode*/) { value.second = false; },
-                                "",
-                                headers,
-                                secureCommunication);
+                                         // Check if the response has the expected number of columns and if the status
+                                         // is green.
+                                         if (fields.size() == HealthCheckColumns::SIZE
+                                             && fields.at(HealthCheckColumns::STATUS).compare("green") == 0)
+                                         {
+                                             value.second = true;
+                                         }
+                                         else
+                                         {
+                                             value.second = false;
+                                         }
+                                     }
+                                     else
+                                     {
+                                         value.second = false; // LCOV_EXCL_LINE
+                                     }
+                                 },
+                                 [&](const std::string& error, const long /*statusCode*/)
+                                 {
+                                     value.second = false;
+                                 }});
                         }
                     }
                 }
@@ -174,7 +174,7 @@ public:
     bool isAvailable(const std::string& value)
     {
         // Check if the server is available.
-        std::lock_guard<std::mutex> lock(m_mutex);
+        std::scoped_lock lock(m_mutex);
         return m_values.at(value);
     }
 };
